requirements.txt
click==7.0
sqlalchemy==1.3.10
snowflake-connector-python==2.0.3
snowflake-sqlalchemy==1.1.16
pandas==0.25.3
dnspython==1.16.0
